feature: 路由配置新增mobile类型
